+++ fastvideo/models/loader/fsdp_load.py
@@ -210,6 +210,7 @@ def maybe_load_fsdp_model(
 
 def shard_model(
     model,
+    *,
     cpu_offload: bool,
     reshard_after_forward: bool = True,
     mp_policy: MixedPrecisionPolicy | None = MixedPrecisionPolicy(),  # noqa
